separated newoverriderow into its own component; separated some data dictionaries into a data file
