+++ electron/ui/src/components/OverrideTable/OverrideTableRows.js
@@ -1,29 +1,9 @@
 import React from 'react';
 import {useEffect, useState} from 'react';
-import { TableBody, TableCell, TableRow, TextField, Tooltip, Checkbox, Select, MenuItem, FormControl, InputLabel, IconButton } from '@mui/material';
-import CheckCircleIcon from '@mui/icons-material/CheckCircle';
-import CheckCircleOutlineIcon from '@mui/icons-material/CheckCircleOutline';
-import CancelIcon from '@mui/icons-material/Cancel';
+import { TableBody, TableCell, TableRow, TextField, Tooltip, Checkbox, Select, MenuItem, FormControl, InputLabel } from '@mui/material';
+import NewBinaryVariableRow from './NewOverrideRow';
+import { INFRASTRUCTURE_CAPEX_MAPPING }  from '../../assets/InfrastructureCapexMapping'
 import CategoryNames from '../../assets/CategoryNames.json'
-
-const OVERRIDE_PRESET_VALUES = {
-  "Pipeline Construction": {
-    input_table: "PipelineDiameterValues",
-    variable_name: "vb_y_Pipeline_dict"
-  },
-  "Storage Facility": {
-    input_table: "StorageCapacityIncrements",
-    variable_name: "vb_y_Storage_dict"
-  },
-  "Disposal Facility": {
-    input_table: "DisposalCapacityIncrements",
-    variable_name: "vb_y_Disposal_dict"
-  },
-  "Treatment Facility": {
-    input_table: "TreatmentCapacityIncrements",
-    variable_name: "vb_y_Treatment_dict"
-  },
-}
 
 export default function OverrideTableRows(props) {  
   const {
@@ -37,7 +17,7 @@ export default function OverrideTableRows(props) {
         handleCheckOverride,
         handleInputOverrideValue,
         newInfrastructureOverrideRow,
-        setNewInfrastructureOverrideRow
+        setNewInfrastructureOverrideRow,
   } = props
 
   const styles = {
@@ -122,455 +102,6 @@ return (
 );
 }
 
-function NewBinaryVariableRow(props) {  
-  const {
-      category,
-      scenario,
-      handleCheckOverride,
-      handleInputOverrideValue
-  } = props
-
-  const styles ={
-      firstCol: {
-      backgroundColor: "#f4f4f4", 
-      border:"1px solid #ddd",
-      position: 'sticky',
-      left: 0,
-
-      },
-      other: {
-      minWidth: 100,
-      border:"1px solid #ddd"
-      },
-      inputDifference: {
-      backgroundColor: "rgb(255,215,0, 0.5)",
-      minWidth: 100,
-      border:"1px solid #ddd"
-      },
-  }
-  const [ value, setValue ] = useState([])
-  const [ rowName, setRowName ] = useState("")
-  const [ location, setLocation ] = useState("")
-  const [ destination, setDestination ] = useState("")
-  const [ technology, setTechnology ] = useState("")
-  const [ capacity, setCapacity ] = useState("")
-  const [ uniqueIndex, setUniqueIndex ] = useState('')
-  const [ overrideChecked, setOverrideChecked ] = useState(true)
-  const [ presetValues, setPresetValues ] = useState({})
-  const [ complete, setComplete ] = useState(false)
-  const [ nodeConnections, setNodeConnections ] = useState({})
-  const pipelineTables = ['PNA', 'CNA','CCA','NNA', 'NCA','NKA', 'NRA', 'NSA','FCA','RCA','RNA','RSA','SCA','SNA']
-  const pipelineTypes = {'P': 'ProductionPads', 'C': 'CompletionsPads', 'N': "NetworkNodes", 'F': 'FreshwaterSources', 'R': 'TreatmentSites', 'S': 'StorageSites'}
-  
-  useEffect(() => {
-
-    //generate location options for pipeline consruction
-    console.log('use effect in place')
-    try {
-      
-        // TODO: generate the pipeline construction options
-        let connections = []
-        let connectionsDictionary = {}
-        let allNodes = new Set()
-        for (let tableName of pipelineTables) {
-          let table = scenario.data_input.df_parameters[tableName]
-          // let rowNames = table['ProductionPads']
-          let rowNames = table[pipelineTypes[tableName.charAt(0)]]
-          
-          for(let colName of Object.keys(table)) {
-            let row = table[colName]
-            let idx = 0
-            for (let element of row) {
-              if (element === 1 || element === "1") {
-                connections.push([rowNames[idx], colName])
-                if (Object.keys(connectionsDictionary).includes(rowNames[idx])) {
-                  connectionsDictionary[rowNames[idx]].push(colName)
-                } else connectionsDictionary[rowNames[idx]] = [colName]
-                allNodes.add(rowNames[idx])
-                allNodes.add(colName)
-              }
-              idx+=1
-            }
-            
-          }
-        }
-        // console.log(connections)
-        console.log(connectionsDictionary)
-        // console.log(allNodes)
-        setNodeConnections(connectionsDictionary)
-      
-    } catch (e) {
-      console.error(e)
-      // console.log("unable to generate infrastructure buildout options from input table, using generic input")
-      return 
-    }
-  },[scenario, category])
-
-  const getValueSelectValue = () => {
-    if (scenario.override_values[category][uniqueIndex] !== undefined) {
-
-      if(scenario.override_values[category][uniqueIndex].variable === "vb_y_Storage_dict" || scenario.override_values[category][uniqueIndex].variable === "vb_y_Disposal_dict") {
-        if(scenario.override_values[category][uniqueIndex].indexes.length>=2) return scenario.override_values[category][uniqueIndex].indexes[1]
-        else return ""
-      } else {
-        if(scenario.override_values[category][uniqueIndex].indexes.length>=3) return scenario.override_values[category][uniqueIndex].indexes[2]
-        else return ""
-      }
-    } else return ""
-  }
-
-  const handleInput = (event) => {
-    let number_value
-    if (value[0] === "Treatment Facility") {
-      number_value = presetValues[technology][event.target.value]
-    } else {
-      number_value = presetValues[event.target.value]
-    } 
-    if(number_value === 0) handleInputOverrideValue(event, true)
-    else handleInputOverrideValue(event, false)
-  }
-
-  const generateValueOptions = () => {
-    // if (Object.keys(OVERRIDE_PRESET_VALUES).includes(rowName)) {
-      try {
-        if (rowName === "Treatment Facility") {
-          return (
-            Object.entries(presetValues[technology]).map(([key,value]) => (
-              <MenuItem key={`${key}_${value}`} value={key}>
-                {value}
-              </MenuItem>
-            ))
-          ) 
-        }
-        else {
-          return (
-            Object.entries(presetValues).map(([key,value]) => (
-              <MenuItem key={`${key}_${value}`} value={key}>
-                {value}
-              </MenuItem>
-            )) 
-          )
-        }
-        
-        
-      } catch (e) {
-        // console.error(e)
-        // console.log("unable to generate infrastructure buildout options from input table, using generic input")
-        return 
-      }
-    // }
-      
-  }
-
-  const generateLocationOptions = () => {
-    try {
-      if (rowName === "Treatment Facility") {
-        return (
-          scenario.data_input.df_sets.TreatmentSites.map((v,i) => (
-            <MenuItem key={`${i}_${v}`} value={v}>
-              {v}
-            </MenuItem>
-          ))
-        ) 
-      }
-      else if (rowName === "Disposal Facility") {
-        return (
-          scenario.data_input.df_sets.SWDSites.map((v,i) => (
-            <MenuItem key={`${i}_${v}`} value={v}>
-              {v}
-            </MenuItem>
-          ))
-        )
-      }
-      else if (rowName === "Storage Facility") {
-        return (
-          scenario.data_input.df_sets.StorageSites.map((v,i) => (
-            <MenuItem key={`${i}_${v}`} value={v}>
-              {v}
-            </MenuItem>
-          ))
-        )
-      }
-      else if (rowName === "Pipeline Construction") {
-        return (
-          Object.entries(nodeConnections).map(([k,v]) => (
-            <MenuItem key={`${k}_${v}`} value={k}>
-              {k}
-            </MenuItem>
-          ))
-        )
-      }
-      
-      
-    } catch (e) {
-      console.error(e)
-      // console.log("unable to generate infrastructure buildout options from input table, using generic input")
-      return 
-    }
-
-  }
-
-  const generateDestinationOptions = () => {
-    if( location!=="") {
-      return (
-        nodeConnections[location].map((v,i) => (
-          <MenuItem key={`${i}_${v}`} value={v}>
-            {v}
-          </MenuItem>
-        ))
-      )
-    } else return null
-    
-  } 
-
-  const generatePresetValues = (rowName) => {
-    try {
-      let preset_value_table = scenario.data_input.df_parameters[OVERRIDE_PRESET_VALUES[rowName].input_table]
-      let preset_values = {}
-        if(rowName === "Treatment Facility") {
-          let technologyNamesKey = "TreatmentCapacities"
-          let technologies = scenario.data_input.df_parameters[OVERRIDE_PRESET_VALUES[rowName].input_table][technologyNamesKey]
-          let len = technologies.length
-          for (let i = 0; i < len; i++) {
-            let each = technologies[i]
-            preset_values[each] = {}
-            for (let row of Object.keys(preset_value_table)) {
-              if (row !== technologyNamesKey) {
-                preset_values[each][row] = preset_value_table[row][i]
-              }
-            }
-          }
-          // setTechnology(tempTechnology)
-        } else {
-  
-          // this is janky but not sure how else to standardize reading the data from these different tables
-          for (let key of Object.keys(preset_value_table)) {
-            if(key !== "VALUE") {
-              let leng = preset_value_table[key].length
-              for(let i = 0; i < leng; i++) {
-                preset_values[preset_value_table[key][i]] = preset_value_table.VALUE[i]
-              }
-            }
-          }
-        }
-        console.log('preset values')
-        console.log(preset_values)
-        setPresetValues(preset_values)
-    } catch(e) {
-      // console.log('failed to generated override table rows: ')
-      // console.log(e)
-    }
-
-  }
-
-  // const generateCapacityOptions = () => {
-
-  // }
-
-
-  const handleSelectRowName = (event) => {
-    let row_name = event.target.value
-    setRowName(row_name)
-    setCapacity("")
-    setLocation("")
-    setDestination("")
-    setTechnology("")
-    // setComplete(false)
-    generatePresetValues(row_name)
-  }
-
-  const handleSelectLocation = (event) => {
-    setLocation(event.target.value)
-  }
-
-  const handleSelectDestination = (event) => {
-    setDestination(event.target.value)
-  }
-
-  const handleSelectTechnology = (event) => {
-    // console.log('selected '+event.target.value)
-    // let tempValue = [...value]
-    // tempValue[5] = event.target.value
-    // setValue(tempValue)
-    setTechnology(event.target.value)
-    // handleInputOverrideValue(event)
-  }
-
-  const handleSelectCapacity = (event) => {
-    setCapacity(event.target.value)
-  }
-
-  const checkForCompletion = () => {
-    if (rowName === "") return true
-    else if (rowName === "Treatment Facility") {
-      if (location === "" || technology === "" || capacity === "") return true
-      else return false
-    }
-    else if (rowName === "Disposal Facility") {
-      if (location === "" || capacity === "") return true
-      else return false
-    }
-    else if (rowName === "Storage Facility") {
-      if (location === "" || capacity === "") return true
-      else return false
-    }
-    else if (rowName === "Pipeline Construction") {
-      if (location === "" || capacity === "" || destination === "") return true
-      else return false
-    }
-  }
-
-
-return (
-      <TableRow>
-            <TableCell 
-              align={"left"} 
-              style={styles.firstCol}>
-                  <FormControl sx={{ width: "100%" }} size="small">
-                  <InputLabel id="">CAPEX Type</InputLabel>
-                  <Select
-                    labelId=""
-                    id="rowname"
-                    name={`rowname`}
-                    value={rowName}
-                    label="CAPEX Type"
-                    onChange={handleSelectRowName}
-                    autoFocus
-                  >
-                    {
-                      ["Treatment Facility", "Disposal Facility", "Storage Facility", "Pipeline Construction"].map((key,idx) => (
-                        <MenuItem key={`${key}_${idx}`} value={key}>
-                          {key}
-                        </MenuItem>
-                      )) 
-                    }
-                  </Select>
-                </FormControl>
-            </TableCell>
-            <TableCell 
-              align={"left"} 
-              style={styles.other}>
-                  <FormControl sx={{ width: "100%" }} size="small">
-                  <InputLabel id="">Origin</InputLabel>
-                  <Select
-                    disabled={rowName===""}
-                    labelId=""
-                    id="location"
-                    name={`location`}
-                    value={location}
-                    label="Origin"
-                    onChange={handleSelectLocation}
-                    MenuProps={{ PaperProps: { sx: { maxHeight: 200 } } }}
-                  >
-                    {
-                      generateLocationOptions()
-                    }
-                  </Select>
-                </FormControl>
-            </TableCell>
-            <TableCell 
-              align={"left"} 
-              style={styles.other}>
-                {rowName === "Pipeline Construction" ? 
-                  <FormControl sx={{ width: "100%" }} size="small">
-                  <InputLabel id="">Destination</InputLabel>
-                  <Select
-                    disabled={location===""}
-                    labelId=""
-                    id="destination"
-                    name={`destination`}
-                    value={destination}
-                    label="Destination"
-                    onChange={handleSelectDestination}
-                  >
-                    {
-                      generateDestinationOptions()
-                    }
-                  </Select>
-                </FormControl>
-                :
-                "--"
-              }
-            </TableCell>
-            <TableCell 
-              align={"left"} 
-              style={styles.other}>
-                {rowName === "Treatment Facility" ? 
-                  <FormControl sx={{ width: "100%" }} size="small">
-                  <InputLabel id="">Technology</InputLabel>
-                  <Select
-                    labelId=""
-                    id="technology"
-                    name={`technology`}
-                    value={technology}
-                    label="Technology"
-                    onChange={handleSelectTechnology}
-                  >
-                    {
-                      Object.keys(presetValues).map((key,idx) => (
-                        <MenuItem key={`${key}_${idx}`} value={key}>
-                          {key}
-                        </MenuItem>
-                      )) 
-                    }
-                  </Select>
-                </FormControl>
-                :
-                "--"
-              }
-            </TableCell>
-            <TableCell 
-              align={"left"} 
-              style={styles.other}>
-                  <FormControl sx={{ width: "100%" }} size="small">
-                  <InputLabel id="">Capacity</InputLabel>
-                  <Select
-                    disabled={rowName==="" || (rowName==="Treatment Facility" && technology==="")}
-                    labelId=""
-                    id="capacity"
-                    name={`capacity`}
-                    value={capacity}
-                    label="Capacity"
-                    onChange={handleSelectCapacity}
-                  >
-                    {
-                      generateValueOptions()
-                    }
-                  </Select>
-                </FormControl>
-            </TableCell>
-            <TableCell 
-              align={"left"} 
-              style={styles.other}>
-                  {
-                    rowName === "Pipeline Construction" ? "in" : 
-                    rowName === "Storage Facility" ? "bbl" : 
-                    ["Disposal Facility", "Treatment Facility"].includes(rowName) ? "bbl/d" : 
-                    "--"
-                   }
-            </TableCell>
-            <TableCell 
-              align="center"
-              style={styles.other}>
-                <Checkbox
-                    checked={overrideChecked}
-                    // onChange={() => handleCheckOverride(uniqueIndex, displayValue)}
-                />
-            </TableCell>
-            <TableCell 
-              align={"center"} 
-              style={styles.other}>
-                  <IconButton color="success" sx={{marginRight:"10px"}} disabled={checkForCompletion()}>
-                    <CheckCircleIcon />
-                  </IconButton>
-                  <IconButton color="error" sx={{marginLeft:"10px"}} disabled={checkForCompletion()}>
-                    <CancelIcon/>
-                  </IconButton>
-            </TableCell>
-        </TableRow>
-    )
-}
-
 function BinaryVariableRow(props) {  
   const {
       category, 
@@ -615,7 +146,7 @@ function BinaryVariableRow(props) {
     setDisplayValue(tempDisplayValue)
 
     try {
-      let preset_value_table = scenario.data_input.df_parameters[OVERRIDE_PRESET_VALUES[tempDisplayValue[0]].input_table]
+      let preset_value_table = scenario.data_input.df_parameters[INFRASTRUCTURE_CAPEX_MAPPING[tempDisplayValue[0]].input_table]
       let preset_values = {}
         if(tempDisplayValue[0] === "Treatment Facility") {
           // check for technology in override values. if found, use that. else, use the value from scenario
@@ -627,7 +158,7 @@ function BinaryVariableRow(props) {
           }
            
           let technologyNamesKey = "TreatmentCapacities"
-          let technologies = scenario.data_input.df_parameters[OVERRIDE_PRESET_VALUES[tempDisplayValue[0]].input_table][technologyNamesKey]
+          let technologies = scenario.data_input.df_parameters[INFRASTRUCTURE_CAPEX_MAPPING[tempDisplayValue[0]].input_table][technologyNamesKey]
           let len = technologies.length
           for (let i = 0; i < len; i++) {
             let each = technologies[i]
@@ -697,11 +228,11 @@ function BinaryVariableRow(props) {
   }
 
   const generateInfrastructureBuildoutValueOptions = (value, index) => {
-    if (Object.keys(OVERRIDE_PRESET_VALUES).includes(value[0])) {
+    if (Object.keys(INFRASTRUCTURE_CAPEX_MAPPING).includes(value[0])) {
       try {
         return (
           <Tooltip 
-            title={Object.keys(scenario.override_values[category]).includes(""+index) ? `To add more options, edit the ${CategoryNames[OVERRIDE_PRESET_VALUES[value[0]].input_table]} table in the data input section.` : ''} 
+            title={Object.keys(scenario.override_values[category]).includes(""+index) ? `To add more options, edit the ${CategoryNames[INFRASTRUCTURE_CAPEX_MAPPING[value[0]].input_table]} table in the data input section.` : ''} 
             placement="top" 
             enterDelay={500}
           >
